Refactor: 패키지 변경 lotto.auto -> lotto
