+++ src/main/java/lotto/Application.java
@@ -1,4 +1,4 @@
-package lotto.auto;
+package lotto;
 
 public class Application {
     public static void main(String[] args) {
