+++ src/main/java/lotto/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.auto;
+package lotto;
 
 import java.util.ArrayList;
 import java.util.List;
