+++ src/main/java/lotto/common/Constants.java
@@ -1,4 +1,4 @@
-package lotto.auto.common;
+package lotto.common;
 
 public class Constants {
     public static final String DEFAULT_PRICE = "1000";
