+++ src/main/java/lotto/ui/InputView.java
@@ -1,4 +1,4 @@
-package lotto.auto.ui;
+package lotto.ui;
 
 import java.util.Scanner;
 
@@ -12,6 +12,11 @@ public class InputView {
     }
 
     public String getInputWinningNumbers() {
+        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
+        return consoleInput();
+    }
+
+    public String getInputBonusNumbers() {
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
         return consoleInput();
     }
