+++ src/test/java/lotto/BuyerTest.java
@@ -1,4 +1,4 @@
-package lotto.auto;
+package lotto;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
