+++ src/test/java/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.auto;
+package lotto;
 
 import org.junit.jupiter.api.Test;
