+++ src/test/java/lotto/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.auto;
+package lotto;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
