+++ src/test/java/lotto/WinningMatcherTest.java
@@ -1,5 +1,6 @@
-package lotto.auto;
+package lotto;
 
+import lotto.common.Constants;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -8,12 +9,12 @@ import java.io.OutputStream;
 import java.io.PrintStream;
 import java.text.NumberFormat;
 
-import static lotto.auto.common.Constants.DECIMAL_POINT;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class WinningMatcherTest {
     private static final String ONCE = "1000";
     private static final int ZERO = 0;
+    private static final String BONUS_NUMBER = "1";
 
     private OutputStream captor;
     Lottos lottos;
@@ -32,11 +33,11 @@ class WinningMatcherTest {
         // given
         Buyer buyer = new Buyer(lottos);
         // when
-        WinningMatcher winningMatcher = new WinningMatcher(buyer, new LottoNumbers("1,2,3,4,5,6"));
+        WinningMatcher winningMatcher = new WinningMatcher(buyer, new LottoNumbers("1,2,3,4,5,6"), BONUS_NUMBER);
         winningMatcher.printProfit(ONCE);
         // then
         float expected = Rank.FIRST.getWinningMoney() / Integer.parseInt(ONCE);
-        assertThat(output()).contains(cutDecimal(DECIMAL_POINT, expected));
+        assertThat(output()).contains(cutDecimal(Constants.DECIMAL_POINT, expected));
     }
 
     @Test
@@ -44,11 +45,11 @@ class WinningMatcherTest {
         // given
         Buyer buyer = new Buyer(lottos);
         // when
-        WinningMatcher winningMatcher = new WinningMatcher(buyer, new LottoNumbers("1,2,3,4,7,8"));
+        WinningMatcher winningMatcher = new WinningMatcher(buyer, new LottoNumbers("1,2,3,4,7,8"), BONUS_NUMBER);
         winningMatcher.printProfit(ONCE);
         // then
         float expected = Rank.THIRD.getWinningMoney() / Integer.parseInt(ONCE);
-        assertThat(output()).contains(cutDecimal(DECIMAL_POINT, expected));
+        assertThat(output()).contains(cutDecimal(Constants.DECIMAL_POINT, expected));
     }
 
     @Test
@@ -56,11 +57,11 @@ class WinningMatcherTest {
         // given
         Buyer buyer = new Buyer(lottos);
         // when
-        WinningMatcher winningMatcher = new WinningMatcher(buyer, new LottoNumbers("12,11,10,9,7,8"));
+        WinningMatcher winningMatcher = new WinningMatcher(buyer, new LottoNumbers("12,11,10,9,7,8"), BONUS_NUMBER);
         winningMatcher.printProfit(ONCE);
         // then
         float expected = ZERO;
-        assertThat(output()).contains(cutDecimal(DECIMAL_POINT, expected));
+        assertThat(output()).contains(cutDecimal(Constants.DECIMAL_POINT, expected));
     }
 
     private final String output() {
