+++ src/test/java/lotto/common/LottoAutoUtilsTest.java
@@ -1,4 +1,4 @@
-package lotto.auto.common;
+package lotto.common;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
